requirements.txt
# WildMatch Requirements

# Core dependencies
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Wikipedia API
wikipedia-api>=0.6.0

# Data science and ML
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
